Fixed
Unable to update profile text has to be changed when user inputs invalid
username and not allowing balance transfer while user is in outgoing
call.

+++ app/src/main/java/com/yo/android/ui/TransferBalanceActivity.java
@@ -25,6 +25,8 @@ import com.yo.android.helpers.RegisteredContactsViewHolder;
 import com.yo.android.helpers.Settings;
 import com.yo.android.photo.TextDrawable;
 import com.yo.android.photo.util.ColorGenerator;
+import com.yo.android.pjsip.MyCall;
+import com.yo.android.pjsip.YoSipService;
 import com.yo.android.util.Constants;
 import com.yo.android.util.Util;
 import com.yo.android.vox.BalanceHelper;
@@ -146,8 +148,10 @@ public class TransferBalanceActivity extends BaseActivity {
                 if (!TextUtils.isEmpty(amount.trim())) {
                     double val = Double.parseDouble(amount.trim());
                     if (val != 0) {
-                        if (Integer.parseInt(mBalanceHelper.getCurrentBalance()) >= Double.parseDouble(amount)) {
+                        if (Double.parseDouble(mBalanceHelper.getCurrentBalance()) >= Double.parseDouble(amount)) {
+
                             showMessageDialog(amount, phoneNo);
+
                         } else {
                             mToastFactory.showToast(R.string.insufficient_amount);
                         }
